fix: rename dir and package to ab-cli
