+++ ab-cli/developper/root.go
@@ -1,8 +1,8 @@
 package developper
 
 import (
-	"github.com/atomic-blend/backend/cli/developper/test"
-	"github.com/atomic-blend/backend/cli/developper/testall"
+	"github.com/atomic-blend/backend/ab-cli/developper/test"
+	"github.com/atomic-blend/backend/ab-cli/developper/testall"
 	"github.com/spf13/cobra"
 )
