+++ ab-cli/root_cmd/root.go
@@ -9,10 +9,10 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/atomic-blend/backend/cli/config"
-	"github.com/atomic-blend/backend/cli/developper"
-	selfhost "github.com/atomic-blend/backend/cli/self-host"
-	envmapper "github.com/atomic-blend/backend/cli/utils/viperutils"
+	"github.com/atomic-blend/backend/ab-cli/config"
+	"github.com/atomic-blend/backend/ab-cli/developper"
+	selfhost "github.com/atomic-blend/backend/ab-cli/self-host"
+	envmapper "github.com/atomic-blend/backend/ab-cli/utils/viperutils"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
