+++ ab-cli/self-host/init_cmd/root.go
@@ -8,14 +8,14 @@ import (
 	"path"
 	"strings"
 
-	"github.com/atomic-blend/backend/cli/config"
-	bulkfiledownloader "github.com/atomic-blend/backend/cli/ui/bulk_file_downloader"
-	channelselector "github.com/atomic-blend/backend/cli/ui/channel_selector"
-	platformcomponentupdater "github.com/atomic-blend/backend/cli/ui/platform_component_updater"
-	filetypes "github.com/atomic-blend/backend/cli/ui/types/file_types"
-	envfilesutils "github.com/atomic-blend/backend/cli/utils/env_files_utils"
-	envmapper "github.com/atomic-blend/backend/cli/utils/viperutils"
-	"github.com/atomic-blend/backend/cli/utils/yamlutils"
+	"github.com/atomic-blend/backend/ab-cli/config"
+	bulkfiledownloader "github.com/atomic-blend/backend/ab-cli/ui/bulk_file_downloader"
+	channelselector "github.com/atomic-blend/backend/ab-cli/ui/channel_selector"
+	platformcomponentupdater "github.com/atomic-blend/backend/ab-cli/ui/platform_component_updater"
+	filetypes "github.com/atomic-blend/backend/ab-cli/ui/types/file_types"
+	envfilesutils "github.com/atomic-blend/backend/ab-cli/utils/env_files_utils"
+	envmapper "github.com/atomic-blend/backend/ab-cli/utils/viperutils"
+	"github.com/atomic-blend/backend/ab-cli/utils/yamlutils"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
