+++ go.work
@@ -2,7 +2,7 @@ go 1.24.5
 
 use (
 	./auth
-	./cli
+	./ab-cli
 	./grpc
 	./mail
 	./mail-server
